Create Sphinx_with_Markdown.rst
Base .rst to convert to a GitHub Gist for information about using Markdown with Sphinx to output HTML via Recommonmark with Autostructify plus sphinx-markdown-tables and PDF via Pandoc to work around latexpdf not parsing Markdown for PDF in Sphinx. Works around the use of tables in Markdown so they get formatted correctly in both HTML and PDF output so developers can contribute Markdown in GitHub for an automated documentation production system that outputs HTML and PDF via Sphinx in Jenkins builds.