Updated README.md
Added a brief description of what's currently committed to the public_Sphinx_repository stating the purpose and use of the how to notes. Also added a note about needing to share a related build.sh bash script for anyone wanting to automate Sphinx docs builds with Jenkins. This is only the base .rst for review before it gets edited and published as a public GitHub Gist and shared with the open source community. Can also share it with TechCommNZ (https://techcomm.nz/) as a local professional body community for feedback & update and also get education use and feedback & update from Ara GCID (Graduate Certificate of Information Design) course staff & students (https://www.ara.ac.nz/study-options/qualifications-and-courses/programmes/Graduate-Certificate-in-Information-Design-Level-7.xml).